requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
pytest
tenacity
openai
